changes: 1. code refactor TODOs: 1. unit tests; 2. e2e tests
