+++ src/widgets/RepaymentCalculator/components/CalculatorResult/helpers.ts
@@ -1,0 +1,6 @@
+const numberWithCommas = (num: number) => {
+  return num.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')
+}
+
+export const displayRepayment = (repayment: number) =>
+  `$${numberWithCommas(-Math.round(repayment))}`
